+++ klin/ver-nodejs/app/src/commonMain/kotlin/org/opengamestudio/const.kt
@@ -1,0 +1,35 @@
+/**
+ * This file is a part of Kotlin dialect:
+ *     https://github.com/OGStudio/kotlin-dialect
+ * License: CC0
+ * Version: 2.0.0
+ */
+
+package org.opengamestudio
+
+const val ARGUMENT_DBG = "--dbg"
+const val ARGUMENT_FILE = "--file"
+const val ARGUMENT_OUT = "--out"
+const val DBG_LEN = 50
+const val DICTIONARY_DELIMITER = ": "
+const val FIELD_DELIMITER = ": "
+const val FILE_EXTENSION_JAVASCRIPT = "js"
+const val FILE_EXTENSION_KOTLIN = "kt"
+const val FILE_EXTENSION_SWIFT = "swift"
+const val LANGUAGE_JAVASCRIPT = "JavaScript"
+const val LANGUAGE_KOTLIN = "Kotlin"
+const val LANGUAGE_SWIFT = "Swift"
+const val NEWLINE = "\n"
+const val POSTFIX_ENTITY = ":"
+const val PREFIX_ENTITY_COMMENT = "# "
+const val PREFIX_FIELD = "        "
+const val PREFIX_FIELD_COMMENT = "        # "
+const val PREFIX_KOTLIN_ENTITY_COMMENT = "// "
+const val PREFIX_KOTLIN_FIELD_COMMENT = "    // "
+const val PREFIX_PREFIX_KOTLIN = "    prefix-kotlin: "
+const val PREFIX_RAW_JAVASCRIPT = "javascript: "
+const val PREFIX_RAW_KOTLIN = "kotlin: "
+const val PREFIX_RAW_SWIFT = "swift: "
+const val PREFIX_TYPE = "    type: "
+const val SECTION_FIELDS = "    fields:"
+const val TYPE_CONTEXT = "context"