+++ src/KDContext.kt
@@ -1,0 +1,44 @@
+/**
+ * This file is a part of Kotlin dialect:
+ *     https://github.com/OGStudio/kotlin-dialect
+ * License: CC0
+ * Version: 2.0.0
+ */
+
+package org.opengamestudio
+
+import kotlin.js.JsExport
+
+@JsExport
+interface KDContext {
+    /**
+     * Name of the field that has just been changed
+     *
+     * Allows should-functions (reducers) to react only to
+     * relevant changes and ignore other changes of KDContext
+     */
+    var recentField: String
+
+    /**
+     * Get field's value by its name
+     */
+    fun <T> field(name: String): T
+    /**
+     * Erase type
+     *
+     * Used by KDController to assign recent field's value
+     */
+    fun fieldAny(name: String): Any {
+        return field(name)
+    }
+    /**
+     * Create a copy of the KDContext derivative
+     *
+     * Used by KDController to treat all derived contexts as KDContext
+     */
+    fun selfCopy(): KDContext
+    /**
+     * Set field's value by its name
+     */
+    fun setField(name: String, value: Any?)
+}