+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # [2.0](https://github.com/OGStudio/kotlin-dialect/pull/1)
 
 1. Replicate [Cross-language dialect](https://github.com/OGStudio/cross-language-dialect)'s functionality solely for Kotlin
-1. Replace JVM translator with Node.js one for easier installation across platforms
+1. Replace JVM translator with Node.js one (now called `klin`) for easier installation across platforms